refactor(core): expose core without react as entrypoint
