+++ packages/fiber/package.json
@@ -35,7 +35,8 @@
   "preconstruct": {
     "entrypoints": [
       "index.tsx",
-      "native.tsx"
+      "native.tsx",
+      "core/index.ts"
     ]
   },
   "scripts": {
