+++ packages/fiber/src/core/events.ts
@@ -1,9 +1,7 @@
 import * as THREE from 'three'
-import { ContinuousEventPriority, DiscreteEventPriority, DefaultEventPriority } from 'react-reconciler/constants'
-import { getRootState } from './utils'
-import type { UseBoundStore } from 'zustand'
+import { Camera, getRootState } from './utils'
 import type { Instance } from './renderer'
-import type { RootState } from './store'
+import type { RootState, Store } from './store'
 
 export interface Intersection extends THREE.Intersection {
   /** The event source (the object which registered the handler) */
@@ -33,7 +31,6 @@ export interface IntersectionEvent<TSourceEvent> extends Intersection {
   stopped: boolean
 }
 
-export type Camera = THREE.OrthographicCamera | THREE.PerspectiveCamera
 export type ThreeEvent<TEvent> = IntersectionEvent<TEvent>
 export type DomEvent = PointerEvent | MouseEvent | WheelEvent
 
@@ -97,30 +94,6 @@ function makeId(event: Intersection) {
   return (event.eventObject || event.object).uuid + '/' + event.index + event.instanceId
 }
 
-// https://github.com/facebook/react/tree/main/packages/react-reconciler#getcurrenteventpriority
-// Gives React a clue as to how import the current interaction is
-export function getEventPriority() {
-  let name = window?.event?.type
-  switch (name) {
-    case 'click':
-    case 'contextmenu':
-    case 'dblclick':
-    case 'pointercancel':
-    case 'pointerdown':
-    case 'pointerup':
-      return DiscreteEventPriority
-    case 'pointermove':
-    case 'pointerout':
-    case 'pointerover':
-    case 'pointerenter':
-    case 'pointerleave':
-    case 'wheel':
-      return ContinuousEventPriority
-    default:
-      return DefaultEventPriority
-  }
-}
-
 /**
  * Release pointer captures.
  * This is called by releasePointerCapture in the API, and when an object is removed.
@@ -142,7 +115,7 @@ function releaseInternalPointerCapture(
   }
 }
 
-export function removeInteractivity(store: UseBoundStore<RootState>, object: THREE.Object3D) {
+export function removeInteractivity(store: Store, object: THREE.Object3D) {
   const { internal } = store.getState()
   // Removes every trace of an object from the data store
   internal.interaction = internal.interaction.filter((o) => o !== object)
@@ -158,7 +131,7 @@ export function removeInteractivity(store: UseBoundStore<RootState>, object: THR
   })
 }
 
-export function createEvents(store: UseBoundStore<RootState>) {
+export function createEvents(store: Store) {
   const temp = new THREE.Vector3()
 
   /** Calculates delta */
