+++ packages/fiber/src/core/store.ts
@@ -1,6 +1,5 @@
 import * as THREE from 'three'
-import * as React from 'react'
-import create, { GetState, SetState, StoreApi, UseBoundStore } from 'zustand'
+import create, { GetState, SetState, StoreApi } from 'zustand/vanilla'
 import { prepare } from './renderer'
 import { DomEvent, EventManager, PointerCaptureTarget, ThreeEvent } from './events'
 import { calculateDpr, Camera, isOrthographicCamera, updateCamera } from './utils'
@@ -21,14 +20,10 @@ export const privateKeys = [
 
 export type PrivateKeys = typeof privateKeys[number]
 
-export interface Intersection extends THREE.Intersection {
-  eventObject: THREE.Object3D
-}
-
 export type Subscription = {
   ref: React.MutableRefObject<RenderCallback>
   priority: number
-  store: UseBoundStore<RootState, StoreApi<RootState>>
+  store: Store
 }
 
 export type Dpr = number | [min: number, max: number]
@@ -75,11 +70,7 @@ export type InternalState = {
   capturedMap: Map<number, Map<THREE.Object3D, PointerCaptureTarget>>
   initialClick: [x: number, y: number]
   initialHits: THREE.Object3D[]
-  subscribe: (
-    callback: React.MutableRefObject<RenderCallback>,
-    priority: number,
-    store: UseBoundStore<RootState, StoreApi<RootState>>,
-  ) => () => void
+  subscribe: (callback: React.MutableRefObject<RenderCallback>, priority: number, store: Store) => () => void
 }
 
 export type RootState = {
@@ -146,17 +137,17 @@ export type RootState = {
   /** When the canvas was clicked but nothing was hit */
   onPointerMissed?: (event: MouseEvent) => void
   /** If this state model is layerd (via createPortal) then this contains the previous layer */
-  previousRoot?: UseBoundStore<RootState, StoreApi<RootState>>
+  previousRoot?: Store
   /** Internals */
   internal: InternalState
 }
 
-const context = React.createContext<UseBoundStore<RootState>>(null!)
+export type Store = StoreApi<RootState>
 
 const createStore = (
   invalidate: (state?: RootState, frames?: number) => void,
   advance: (timestamp: number, runGlobalEffects?: boolean, state?: RootState, frame?: THREE.XRFrame) => void,
-): UseBoundStore<RootState> => {
+): Store => {
   const rootState = create<RootState>((set, get) => {
     const position = new THREE.Vector3()
     const defaultTarget = new THREE.Vector3()
@@ -278,7 +269,7 @@ const createStore = (
         active: false,
         priority: 0,
         frames: 0,
-        lastEvent: React.createRef(),
+        lastEvent: { current: null },
 
         interaction: [],
         hovered: new Map<string, ThreeEvent<DomEvent>>(),
@@ -287,11 +278,7 @@ const createStore = (
         initialHits: [],
         capturedMap: new Map(),
 
-        subscribe: (
-          ref: React.MutableRefObject<RenderCallback>,
-          priority: number,
-          store: UseBoundStore<RootState, StoreApi<RootState>>,
-        ) => {
+        subscribe: (ref: React.MutableRefObject<RenderCallback>, priority: number, store: Store) => {
           const internal = get().internal
           // If this subscription was given a priority, it takes rendering into its own hands
           // For that reason we switch off automatic rendering and increase the manual flag
@@ -351,4 +338,4 @@ const createStore = (
   return rootState
 }
 
-export { createStore, context }
+export { createStore }
