+++ packages/fiber/src/index.tsx
@@ -2,19 +2,10 @@ export * from './three-types'
 import * as ReactThreeFiber from './three-types'
 export { ReactThreeFiber }
 export type { BaseInstance, LocalState } from './core/renderer'
-export type {
-  Intersection,
-  Subscription,
-  Dpr,
-  Size,
-  Viewport,
-  RenderCallback,
-  Performance,
-  RootState,
-} from './core/store'
-export type { ThreeEvent, Events, EventManager, ComputeFunction } from './core/events'
+export type { Subscription, Dpr, Size, Viewport, RenderCallback, Performance, RootState } from './core/store'
+export type { ThreeEvent, Events, EventManager, ComputeFunction, Intersection } from './core/events'
 export type { ObjectMap, Camera } from './core/utils'
 export * from './web/Canvas'
 export { createEvents } from './core/events'
 export { createPointerEvents as events } from './web/events'
-export * from './core'
+export * from './react'
