+++ packages/fiber/src/native.tsx
@@ -2,18 +2,9 @@ export * from './three-types'
 import * as ReactThreeFiber from './three-types'
 export { ReactThreeFiber }
 export type { BaseInstance, LocalState } from './core/renderer'
-export type {
-  Intersection,
-  Subscription,
-  Dpr,
-  Size,
-  Viewport,
-  RenderCallback,
-  Performance,
-  RootState,
-} from './core/store'
-export type { ThreeEvent, Events, EventManager, ComputeFunction } from './core/events'
+export type { Subscription, Dpr, Size, Viewport, RenderCallback, Performance, RootState } from './core/store'
+export type { ThreeEvent, Events, EventManager, ComputeFunction, Intersection } from './core/events'
 export type { ObjectMap, Camera } from './core/utils'
 export * from './native/Canvas'
 export { createTouchEvents as events } from './native/events'
-export * from './core'
+export * from './react'
