+++ packages/fiber/src/native/Canvas.tsx
@@ -2,8 +2,9 @@ import * as React from 'react'
 import * as THREE from 'three'
 import { View, ViewProps, ViewStyle, LayoutChangeEvent, StyleSheet, PixelRatio } from 'react-native'
 import { ExpoWebGLRenderingContext, GLView } from 'expo-gl'
-import { SetBlock, Block, ErrorBoundary, useMutableCallback } from '../core/utils'
-import { extend, createRoot, unmountComponentAtNode, RenderProps, ReconcilerRoot } from '../core'
+import { SetBlock, Block, ErrorBoundary } from '../react/utils'
+import { useMutableCallback } from '../react/hooks'
+import { extend, createRoot, unmountComponentAtNode, RenderProps, ReconcilerRoot } from '../react'
 import { createTouchEvents } from './events'
 import { RootState, Size } from '../core/store'
 import { polyfills } from './polyfills'
