+++ packages/fiber/src/react/index.tsx
@@ -8,32 +8,24 @@ import {
   Renderer,
   createStore,
   isRenderer,
-  context,
   RootState,
   Size,
   Dpr,
   Performance,
   PrivateKeys,
   privateKeys,
-} from './store'
-import { createRenderer, extend, Root } from './renderer'
-import { createLoop, addEffect, addAfterEffect, addTail } from './loop'
-import { getEventPriority, EventManager, ComputeFunction } from './events'
-import {
-  is,
-  dispose,
-  calculateDpr,
-  EquConfig,
-  getRootState,
-  useIsomorphicLayoutEffect,
-  Camera,
-  updateCamera,
-  setDeep,
-} from './utils'
-import { useStore } from './hooks'
+} from '../core/store'
+import { extend, Root } from '../core/renderer'
+import { createRenderer, ReactThreeRoot } from './reconciler'
+import { createLoop, addEffect, addAfterEffect, addTail } from '../core/loop'
+import { EventManager, ComputeFunction } from '../core/events'
+import { is, dispose, calculateDpr, EquConfig, getRootState, Camera, updateCamera, setDeep } from '../core/utils'
+import { useStore, useIsomorphicLayoutEffect } from './hooks'
 import { OffscreenCanvas } from 'three'
+import { context } from './context'
+import { getEventPriority } from './utils'
 
-const roots = new Map<Element, Root>()
+const roots = new Map<Element, ReactThreeRoot>()
 const { invalidate, advance } = createLoop(roots)
 const { reconciler, applyProps } = createRenderer(roots, getEventPriority)
 const shallowLoose = { objects: 'shallow', strict: false } as EquConfig
@@ -159,7 +151,7 @@ function createRoot<TCanvas extends Element>(canvas: TCanvas): ReconcilerRoot<TC
         console.error
 
   // Create store
-  const store = prevStore || createStore(invalidate, advance)
+  const store = prevStore || create(createStore(invalidate, advance))
   // Create renderer
   const fiber =
     prevFiber || reconciler.createContainer(store, ConcurrentRoot, null, false, null, '', logRecoverableError, null)
