+++ packages/fiber/src/web/Canvas.tsx
@@ -2,8 +2,9 @@ import * as React from 'react'
 import * as THREE from 'three'
 import useMeasure from 'react-use-measure'
 import type { Options as ResizeOptions } from 'react-use-measure'
-import { SetBlock, Block, ErrorBoundary, useMutableCallback, useIsomorphicLayoutEffect } from '../core/utils'
-import { ReconcilerRoot, extend, createRoot, unmountComponentAtNode, RenderProps } from '../core'
+import { SetBlock, Block, ErrorBoundary } from '../react/utils'
+import { useMutableCallback, useIsomorphicLayoutEffect } from '../react/hooks'
+import { ReconcilerRoot, extend, createRoot, unmountComponentAtNode, RenderProps } from '../react'
 import { createPointerEvents } from './events'
 import { DomEvent } from '../core/events'
